requirements.txt
# Dependencias principales
requests>=2.28.0
selenium>=4.10.0
webdriver-manager>=3.8.0
python-dotenv>=0.19.0
beautifulsoup4>=4.11.0
PyYAML>=6.0

# Dependencias de timezone
tzdata>=2022.7

# Para posibles futuras extensiones
pandas>=1.5.0
python-dateutil>=2.8.0
